move file_utils to its own source file
